Fee double로 수정

+++ src/main/java/com/example/finalproject/domain/order/Order.java
@@ -25,7 +25,7 @@ public class Order {
     @OneToOne
     private Delivery delivery; // 배송 고유번호
 
-    private Double totalOrderAmount; // 주문 전체 금액
+    private Integer totalOrderAmount; // 주문 전체 금액
 
     private Double fee; // 주문 전체 수수료
 
@@ -34,7 +34,7 @@ public class Order {
 
 
     @Builder
-    public Order(Integer id, User user, Delivery delivery, Double totalOrderAmount, Double fee, Timestamp orderDate) {
+    public Order(Integer id, User user, Delivery delivery, Integer totalOrderAmount, Double fee, Timestamp orderDate) {
         this.id = id;
         this.user = user;
         this.delivery = delivery;
